+++ client/src/App.vue
@@ -4,10 +4,7 @@
     <AeMain id="main">
       <b-container class="container">
         <div class="container" id="content">
-          <transition
-            name="slide"
-            mode="out-in"
-          >
+          <transition name="slide" mode="out-in">
             <router-view />
           </transition>
         </div>
@@ -18,52 +15,66 @@
 </template>
 
 <script>
-	import { AeMain } from "@aeternity/aepp-components";
+import { AeMain } from "@aeternity/aepp-components";
 
-	import "@aeternity/aepp-components/dist/aepp.global.css";
-	import "@aeternity/aepp-components/dist/aepp.components.css";
-	import "sweetalert2/dist/sweetalert2.min.css";
-	import LoadingScreen from "./components/LoadingScreen";
-	import AeBanner from "./components/AeBanner";
+import "@aeternity/aepp-components/dist/aepp.global.css";
+import "@aeternity/aepp-components/dist/aepp.components.css";
+import "sweetalert2/dist/sweetalert2.min.css";
+import LoadingScreen from "./components/LoadingScreen";
+import AeBanner from "./components/AeBanner";
 
-	export default {
-		name: "App",
-		components: { AeMain, LoadingScreen, AeBanner },
-		data() {
-			return { isLoading: false };
-		},
-		beforeMount() {
-			this.$store.state.route && this.$router.replace({name: this.$store.state.route.name});
-		}
-	};
+export default {
+  name: "App",
+  components: { AeMain, LoadingScreen, AeBanner },
+  data() {
+    return { isLoading: false };
+  },
+  async beforeMount() {
+    // restore channel if available.
+
+	console.log(this.$store.state.channel)
+;    if (this.$store.state.channelOpenDone) {
+      try {
+        console.log("beforeMount(): Channel already opened, reconnecting");
+        //await this.$store.dispatch("openChannel");
+      } catch (err) {
+		  console.error("Cannot re-connect your channel. Reason: " + e.toString());
+	  }
+    }
+
+    // restore route if available.
+    this.$store.state.route && this.$router.replace({ name: this.$store.state.route.name });
+  }
+};
 </script>
 
 <style>
-	.container {
-		height: 80%;
-	}
-	#content {
-		height: 100%;
-		margin-top: 10%;
-		margin-bottom: 10%;
-		border-collapse: collapse;
-	}
-	#app {
-		overflow: hidden;
-		height: 100vh;
-		text-align: center;
-		-webkit-font-smoothing: antialiased;
-		-moz-osx-font-smoothing: grayscale;
-		text-align: center;
-		background-color: #311B58;
-	}
-	#main {
-		height: 100vh;
-		min-height: 0vh;
-		width: 100vw;
-		border-radius:30px 30px 0px 0px;
-	}
-	html, body {
-		overflow: hidden;
-	}
+.container {
+  height: 80%;
+}
+#content {
+  height: 100%;
+  margin-top: 10%;
+  margin-bottom: 10%;
+  border-collapse: collapse;
+}
+#app {
+  overflow: hidden;
+  height: 100vh;
+  text-align: center;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  background-color: #311b58;
+}
+#main {
+  height: 100vh;
+  min-height: 0vh;
+  width: 100vw;
+  border-radius: 30px 30px 0px 0px;
+}
+html,
+body {
+  overflow: hidden;
+}
 </style>